+++ server/models/index.js
@@ -6,7 +6,16 @@ const Sauce = sequelize.define("sauces", {
   image: Sequelize.STRING,
 });
 
+const Items = sequelize.define("items", {
+  name: Sequelize.STRING,
+  description: Sequelize.STRING,
+  price: Sequelize.INTEGER,
+  category: Sequelize.STRING,
+  image: Sequelize.STRING,
+});
+
 module.exports = {
   db: sequelize,
   Sauce,
+  Items
 };
