+++ server/routes/index.js
@@ -2,7 +2,8 @@ const express = require("express");
 const router = express.Router();
 
 // different model routers
-router.use("/sauces", require("./sauces"));
-router.use("/items", require("./items"));
+router.use("/sauces", require("./sauces.js"));
+router.use("/items", require("./items.js"));
+router.use("/products.json", require("./products.js"))
 
 module.exports = router;
